+++ components/collapse/collapseItem.js
@@ -1,10 +1,13 @@
 import React, { PropTypes } from 'react';
 
-const CollapseItem = ({ id, collapsed, name, title, children, type, onChange }) => (
+/**
+ * Collapse item component
+ */
+const CollapseItem = ({ id, expanded, name, title, children, type, onChange }) => (
   <div className="ui-collapse-item">
     <input
       className="ui-collapse-item__input"
-      defaultChecked={collapsed}
+      defaultChecked={expanded}
       id={id}
       name={name}
       onChange={onChange}
@@ -17,12 +20,30 @@ const CollapseItem = ({ id, collapsed, name, title, children, type, onChange }) 
   </div>
 );
 CollapseItem.propTypes = {
+  /**
+   * The collapse should contains CollapseItem components
+   */
   children: PropTypes.node.isRequired,
-  id: PropTypes.string.isRequired,
+  /**
+   * Unique id for input and label Unique id for input and label.
+   */
+  id: PropTypes.string,
+  /**
+   * Specify a function that will be called when a user click on CollapseItem
+   */
   onChange: PropTypes.func,
+  /**
+   * The CollapseItem's title
+   */
   title: PropTypes.node.isRequired,
   name: PropTypes.string.isRequired,
-  collapsed: PropTypes.bool,
+  /**
+   * Determine whether a CollapseItem is collapsed or not by default
+   */
+  expanded: PropTypes.bool,
+  /**
+   * Determine type for input, can be one of ('radio', 'checkbox')
+   */
   type: PropTypes.oneOf(['radio', 'checkbox']),
 };
 CollapseItem.defaultProps = {
