+++ components/index.scss
@@ -2,7 +2,7 @@
 @import "./calendar/calendar.scss";
 @import "./global/global.scss";
 @import "./collapse/collapse.scss";
-@import "./collapseItem/collapseItem.scss";
+@import "./collapse/collapseItem.scss";
 @import "./list/list.scss";
 @import "./modal/modal.scss";
 @import "./price/price.scss";
